+++ index.js
@@ -1,13 +1,7 @@
 module.exports = {
   rules: {
     // 'selectors',
-    'text-and-font': require('./src/rules/textAndFonts'),
-    // 'color-and-background',
-    // 'box-model',
-    // 'positioning',
-    // 'lists',
-    // 'tables',
-    // 'html5'
+    'unsupported-css': require('./src/rules/unsupported-css'),
   },
   configs: {
     recommended: {
